🐛 FIX: Routes Name - Fixed routes names and files
