+++ src/index.js
@@ -1,11 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom";
-// Components
+import App from "./routes/App";
 // Redux
 import { Provider } from "react-redux";
 import { createStore, compose, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
-import App from "./routes/App";
 import reducer from "./reducers";
 
 import initialState from "../initialState.json";
